auth_service/users/views.py
from rest_framework import generics

from users.models import CustomUser as User
from users.serializers import RegisterSerializer


class RegisterView(generics.CreateAPIView):
    queryset = User.objects.all()
    serializer_class = RegisterSerializer
