auth_service/requirements.txt
django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0